[#8] refactor: 이넘을 도메인 모델로 이동
- 도메인 모델로 이넘을 정의하고 다른 계층에서 이를 활용하도록 수정함.
- 엔티티는 이넘을 db에 저장할 때 컨버터를 사용하도록.
