+++ src/main/java/com/baedal/review/application/port/dto/StoreReviewSummary.java
@@ -1,6 +1,6 @@
 package com.baedal.review.application.port.dto;
 
-import com.baedal.review.adapter.persistence.entity.ReviewScore;
+import com.baedal.review.domain.model.ReviewScore;
 import lombok.Builder;
 import lombok.Getter;
